+++ stickers.go
@@ -66,13 +66,13 @@ type (
 		Point string `json:"point"`
 
 		// Shift by X-axis measured in widths of the mask scaled to the face size, from left to right. For example, choosing -1.0 will place mask just to the left of the default mask position.
-		XShift float32 `json:"x_shift"`
+		XShift float64 `json:"x_shift"`
 
 		// Shift by Y-axis measured in heights of the mask scaled to the face size, from top to bottom. For example, 1.0 will place the mask just below the default mask position.
-		YShift float32 `json:"y_shift"`
+		YShift float64 `json:"y_shift"`
 
 		// Mask scaling coefficient. For example, 2.0 means double size.
-		Scale float32 `json:"scale"`
+		Scale float64 `json:"scale"`
 	}
 
 	// SendStickerParameters represents data for SetSticker method.
